frontend: rename files and remove useless code
