+++ frontend/src/components/ui/input.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "../../lib/utils"
+import { cn } from "../../lib/cn.ts"
 
 export function Input({ className, type, ...props }: React.ComponentProps<"input">): React.JSX.Element {
   return (
