+++ frontend/src/components/ui/label.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import { Label as LabelPrimitive } from "radix-ui"
 
-import { cn } from "../../lib/utils"
+import { cn } from "../../lib/cn.ts"
 
 export function Label({ className, ...props }: React.ComponentProps<typeof LabelPrimitive.Root>): React.JSX.Element {
   return (
