+++ frontend/src/components/ui/separator.tsx
@@ -3,7 +3,7 @@
 import * as React from "react"
 import { Separator as SeparatorPrimitive } from "radix-ui"
 
-import { cn } from "../../lib/utils"
+import { cn } from "../../lib/cn.ts"
 
 export function Separator({
   className,
