+++ frontend/src/components/ui/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "../../lib/utils"
+import { cn } from "../../lib/cn.ts"
 
 export function Skeleton({ className, ...props }: React.ComponentProps<"div">): React.JSX.Element {
   return <div data-slot="skeleton" className={cn("animate-pulse rounded-2xl bg-muted", className)} {...props} />
